working on jenkinsfile-buildscript fixes

+++ src/BuildScripts/build.macos.sh
@@ -11,33 +11,26 @@ echo "  ### Starting macOS Build ###"
 
 #if doesn't the packenv directory doesn't exist...
 pushd ..
+rm -rf packenv
 directory="./packenv"
 actfile="./packenv/bin/activate"
-#if [ ! -d "$directory" ]  || [ ! -f "$actfile" ] ; then
-#   python3 -m venv packenv
-#   source packenv/bin/activate
-#fi
-rm -rf packenv
-./pyclean.py
-find . -iname ".pytest_cache" -exec rm -rf {} \;
-sleep 2
-sync
-sync
-sync
-sleep 2
-python3 -m venv packenv
-source packenv bin/activate
+if [ ! -d "$directory" ]  || [ ! -f "$actfile" ] ; then
+   python3 -m venv packenv
+   source packenv/bin/activate
 
-pip install PySide6 PyInstaller
-pip install PyInstaller
-pip install packaging
-pip install psutil
-pip install sphynx  # documentation tool
-pip install myst-parser # supports markdown for sphynx
-pip install requests
-pip install pytest
-pip install astroid
-pip install pylint
+   pip install PySide6 PyInstaller
+   pip install --upgrade PyInstaller pyinstaller-hooks-contrib
+   pip install packaging
+   pip install psutil
+   pip install sphynx  # documentation tool
+   pip install myst-parser # supports markdown for sphynx
+   pip install requests
+   pip install pytest
+   pip install astroid
+   pip install pylint
+else
+   source packenv/bin/activate
+fi
 
 echo "."
 echo "."
